+++ Accounts/supplychain/contracts/src/main/kotlin/net/corda/samples/supplychain/contracts/PaymentStateContract.kt
@@ -1,0 +1,26 @@
+package net.corda.samples.supplychain.contracts
+
+import net.corda.core.contracts.*
+import net.corda.core.transactions.LedgerTransaction
+
+class PaymentStateContract : Contract{
+
+    companion object{
+        const val ID = "net.corda.samples.supplychain.contracts.PaymentStateContract"
+    }
+
+
+    override fun verify(tx: LedgerTransaction) {
+        requireThat {
+            /*
+             *
+             * For the simplicity of the sample, we unconditionally accept all of the transactions.
+             *
+             */
+        }
+    }
+
+    interface Commands : CommandData {
+        class Create : Commands
+    }
+}
