+++ Accounts/supplychain/gradle.properties
@@ -1,4 +1,4 @@
-name=Test
-group=com.template
-version=0.1
+name=Supplychain Cordapp
+group=com.supplychain
+version=1.0
 kotlin.incremental=false
